requirements.txt
pillow>=9.5.0
opencv-python
insightface>=0.7.3
onnxruntime
